+++ transmux/run_polyphase_test.py
@@ -105,7 +105,7 @@ if __name__ == "__main__":
     inds = np.arange(0, block_len * num_channels * num_blocks, dtype=int).reshape(num_blocks, -1)
 
     channelHz = Fs / num_channels
-    rx = PolyphaseRxChannelizer(sample_rate_Hz=Fs, channel_bandwidth_Hz=channelHz, block_size=block_len)
+    rx = PolyphaseRxChannelizer(sample_rate_Hz=Fs, channel_bandwidth_Hz=channelHz)
     print(rx)
 
     # Start the stopwatch / counter
@@ -124,7 +124,7 @@ if __name__ == "__main__":
     print("Elapsed time: {} s".format(t1_stop - t1_start))
     print("Samples per second: {}".format(num_blocks*output.size / (t1_stop - t1_start)))
 
-    tx = PolyphaseTxMultiplexer(sample_rate_Hz=Fs, channel_bandwidth_Hz=channelHz, block_size=block_len)
+    tx = PolyphaseTxMultiplexer(sample_rate_Hz=Fs, channel_bandwidth_Hz=channelHz)
     print(tx)
     inds = np.arange(0, block_len * num_blocks, dtype=int).reshape(num_blocks, -1)
     t1_start = perf_counter()
